📄 feat(debug): add HTML file for Google search results from January 1, 2022
The new HTML file contains the raw output of a Google search conducted on January 1, 2022. This addition serves as a reference for debugging and analysis of search result formats and content from that date.

📸 feat(debug): add Google search screenshot for January 1, 2022

A new screenshot is added to the debug directory to provide a visual reference of the Google search results from that date. This can assist in debugging and verifying search-related functionalities in the application.